+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.31
 
-- ...
+- Fixed markdown writer to preserve titles that start with '!'.
 
 ## 1.0.30